refactor(search): unaccent and Subscriptions component
