+++ src/app/profiles/App/Pages/Subscriptions/Subscriptions.tsx
@@ -1,25 +1,21 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import styled from 'styled-components';
-import {
-  ContributorId,
-  findContributorIn,
-  StatefulContributor
-} from 'libs/domain/contributor';
+import { ContributorId, StatefulContributor } from 'libs/domain/contributor';
 import { Categories } from 'libs/domain/category';
 import { TwoColumns } from 'components/atoms';
 import { SlowerMessageBox } from 'components/molecules/SidebarBox';
+import { Arrow } from 'components/atoms/icons';
+import Search from 'components/molecules/Search/Search';
 import pathToContributor from 'app/profiles/App/pathToContributor';
 import SimilarProfiles from 'app/profiles/App/SimilarProfiles';
-import useContributorsFilters from 'app/profiles/App/useContributorsFilters';
+import useContributorsFilters from 'app/profiles/hooks/useContributorsFilters';
 import {
   ContributorExampleLink,
   ContributorProfileListItem
 } from '../Profiles/List/ProfileList';
 import { Aside, MainCol } from '../Profiles/Profile/Profile';
 import ProfileTabs from '../../ProfileTabs';
-import { Arrow } from '../../../../../components/atoms/icons';
-import Search from '../../../../../components/molecules/Search/Search';
 
 const ContributorsList = styled.div`
   display: grid;
@@ -31,8 +27,7 @@ const TwoColumnsWithMargin = styled(TwoColumns)`
   margin-top: 20px;
 `;
 export interface SubscriptionsProps {
-  subscriptions: StatefulContributor[];
-  contributors: StatefulContributor[];
+  subscribedContributors: StatefulContributor[];
   subscribe: (id: ContributorId) => void;
   unsubscribe: (id: ContributorId) => void;
   className?: string;
@@ -42,8 +37,7 @@ export interface SubscriptionsProps {
 }
 
 const Subscriptions = ({
-  subscriptions,
-  contributors,
+  subscribedContributors,
   subscribe,
   unsubscribe,
   connected,
@@ -51,24 +45,13 @@ const Subscriptions = ({
   categories,
   categoriesLoading
 }: SubscriptionsProps) => {
-  const [initialSubscriptions, setInitialSubscriptions] = useState(
-    subscriptions
-  );
-
-  useEffect(() => {
-    if (initialSubscriptions.length === 0)
-      setInitialSubscriptions(subscriptions);
-  }, [subscriptions]);
   const { t } = useTranslation();
-  const subscriptionsToRender = initialSubscriptions.map(
-    findContributorIn(contributors)
-  );
 
   const [
     filteredSubscriptions,
     setFilter,
     handleChangeSearchContributors
-  ] = useContributorsFilters(subscriptionsToRender);
+  ] = useContributorsFilters(subscribedContributors);
 
   return (
     <>
